database connection

+++ src/resources/boards/board.service.js
@@ -1,13 +1,16 @@
-const boardsRepo = require('./board.memory.repository');
-const tasksRepo = require('../tasks/task.memory.repository');
+const boardsRepo = require('./board.db.repository');
+const tasksRepo = require('../tasks/task.db.repository');
 
 const getAllBoards = () => boardsRepo.getAllBoards();
 const getBoardById = id => boardsRepo.getBoardById(id);
 const createBoard = board => boardsRepo.createBoard(board);
 const updateBoard = board => boardsRepo.updateBoard(board);
 const deleteBoard = async id => {
-  tasksRepo.deleteBoardTasks(id);
   const result = await boardsRepo.deleteBoard(id);
+
+  if (result) {
+    await tasksRepo.deleteBoardTasks(id);
+  }
 
   return result;
 };
